backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==1.2.0
python-dotenv==1.0.0
werkzeug==3.0.0
PyJWT==2.8.0
cloudinary==1.36.0
requests==2.31.0